prefix-selectors: add support for :root pseudo-class
In our modular component CSS, we've been having trouble referencing the root
element of our CSS when used in conjunction with the `prefixSelectors()`
plugin.

A clever solution for this is to re-purpose the `:root` pseudo-class from
CSS3 to mean "the prefixed selector" when encountered, which would give us
a nice way to reference that root element with a clean semantic syntax.